Added game reset
- Game now enters fade in mode after it has faded out from a game over
state
- All variables are reset and a fresh game will be ready to execute

+++ Deluxema/Sound.h
@@ -14,7 +14,7 @@ public:
 private:
 	int id;
 	int volume;
-	int fadeVolume;
+	double fadeVolume;
 };
